requirements.txt
fastapi
uvicorn
streamlit
pandas
sentence-transformers
qdrant-client
hdbscan
numpy
requests
python-dotenv
google-api-python-client
langchain
langchain-huggingface
langchain-groq
langdetect
